+++ src/main/java/com/example/examplemod/ExampleMod.java
@@ -41,33 +41,53 @@ public class ExampleMod {
     public static final String MOD_ID = "examplemod";
     // Directly reference a slf4j logger
     public static final Logger LOGGER = LogUtils.getLogger();
-    // // Create a Deferred Register to hold Blocks which will all be registered under the "examplemod" namespace
-    // public static final DeferredRegister.Blocks BLOCKS = DeferredRegister.createBlocks(MOD_ID);
-    // // Create a Deferred Register to hold Items which will all be registered under the "examplemod" namespace
-    // public static final DeferredRegister.Items ITEMS = DeferredRegister.createItems(MOD_ID);
-    // // Create a Deferred Register to hold CreativeModeTabs which will all be registered under the "examplemod" namespace
-    // public static final DeferredRegister<CreativeModeTab> CREATIVE_MODE_TABS = DeferredRegister.create(Registries.CREATIVE_MODE_TAB, MOD_ID);
+    // // Create a Deferred Register to hold Blocks which will all be registered
+    // under the "examplemod" namespace
+    // public static final DeferredRegister.Blocks BLOCKS =
+    // DeferredRegister.createBlocks(MOD_ID);
+    // // Create a Deferred Register to hold Items which will all be registered
+    // under the "examplemod" namespace
+    // public static final DeferredRegister.Items ITEMS =
+    // DeferredRegister.createItems(MOD_ID);
+    // // Create a Deferred Register to hold CreativeModeTabs which will all be
+    // registered under the "examplemod" namespace
+    // public static final DeferredRegister<CreativeModeTab> CREATIVE_MODE_TABS =
+    // DeferredRegister.create(Registries.CREATIVE_MODE_TAB, MOD_ID);
 
-    // // Creates a new Block with the id "examplemod:example_block", combining the namespace and path
-    // public static final DeferredBlock<Block> EXAMPLE_BLOCK = BLOCKS.registerSimpleBlock("example_block", p -> p.mapColor(MapColor.STONE));
-    // // Creates a new BlockItem with the id "examplemod:example_block", combining the namespace and path
-    // public static final DeferredItem<BlockItem> EXAMPLE_BLOCK_ITEM = ITEMS.registerSimpleBlockItem("example_block", EXAMPLE_BLOCK);
+    // // Creates a new Block with the id "examplemod:example_block", combining the
+    // namespace and path
+    // public static final DeferredBlock<Block> EXAMPLE_BLOCK =
+    // BLOCKS.registerSimpleBlock("example_block", p -> p.mapColor(MapColor.STONE));
+    // // Creates a new BlockItem with the id "examplemod:example_block", combining
+    // the namespace and path
+    // public static final DeferredItem<BlockItem> EXAMPLE_BLOCK_ITEM =
+    // ITEMS.registerSimpleBlockItem("example_block", EXAMPLE_BLOCK);
 
-    // // Creates a new food item with the id "examplemod:example_id", nutrition 1 and saturation 2
-    // public static final DeferredItem<Item> EXAMPLE_ITEM = ITEMS.registerSimpleItem("example_item", p -> p.food(new FoodProperties.Builder()
-    //         .alwaysEdible().nutrition(1).saturationModifier(2f).build()));
+    // // Creates a new food item with the id "examplemod:example_id", nutrition 1
+    // and saturation 2
+    // public static final DeferredItem<Item> EXAMPLE_ITEM =
+    // ITEMS.registerSimpleItem("example_item", p -> p.food(new
+    // FoodProperties.Builder()
+    // .alwaysEdible().nutrition(1).saturationModifier(2f).build()));
 
-    // Creates a creative tab with the id "examplemod:example_tab" for the example item, that is placed after the combat tab
-    // public static final DeferredHolder<CreativeModeTab, CreativeModeTab> EXAMPLE_TAB = CREATIVE_MODE_TABS.register("example_tab", () -> CreativeModeTab.builder()
-    //         .title(Component.translatable("itemGroup.examplemod")) //The language key for the title of your CreativeModeTab
-    //         .withTabsBefore(CreativeModeTabs.COMBAT)
-    //         .icon(() -> EXAMPLE_ITEM.get().getDefaultInstance())
-    //         .displayItems((parameters, output) -> {
-    //             output.accept(EXAMPLE_ITEM.get());// Add the example item to the tab. For your own tabs, this method is preferred over the event
-    //         }).build());
+    // Creates a creative tab with the id "examplemod:example_tab" for the example
+    // item, that is placed after the combat tab
+    // public static final DeferredHolder<CreativeModeTab, CreativeModeTab>
+    // EXAMPLE_TAB = CREATIVE_MODE_TABS.register("example_tab", () ->
+    // CreativeModeTab.builder()
+    // .title(Component.translatable("itemGroup.examplemod")) //The language key for
+    // the title of your CreativeModeTab
+    // .withTabsBefore(CreativeModeTabs.COMBAT)
+    // .icon(() -> EXAMPLE_ITEM.get().getDefaultInstance())
+    // .displayItems((parameters, output) -> {
+    // output.accept(EXAMPLE_ITEM.get());// Add the example item to the tab. For
+    // your own tabs, this method is preferred over the event
+    // }).build());
 
-    // The constructor for the mod class is the first code that is run when your mod is loaded.
-    // FML will recognize some parameter types like IEventBus or ModContainer and pass them in automatically.
+    // The constructor for the mod class is the first code that is run when your mod
+    // is loaded.
+    // FML will recognize some parameter types like IEventBus or ModContainer and
+    // pass them in automatically.
     public ExampleMod(IEventBus modEventBus, ModContainer modContainer) {
         // Register the commonSetup method for modloading
         modEventBus.addListener(this::commonSetup);
@@ -84,14 +104,17 @@ public class ExampleMod {
         ModBlocks.register(modEventBus);
 
         // Register ourselves for server and other game events we are interested in.
-        // Note that this is necessary if and only if we want *this* class (ExampleMod) to respond directly to events.
-        // Do not add this line if there are no @SubscribeEvent-annotated functions in this class, like onServerStarting() below.
+        // Note that this is necessary if and only if we want *this* class (ExampleMod)
+        // to respond directly to events.
+        // Do not add this line if there are no @SubscribeEvent-annotated functions in
+        // this class, like onServerStarting() below.
         NeoForge.EVENT_BUS.register(this);
 
         // Register the item to a creative tab
         modEventBus.addListener(this::addCreative);
 
-        // Register our mod's ModConfigSpec so that FML can create and load the config file for us
+        // Register our mod's ModConfigSpec so that FML can create and load the config
+        // file for us
         modContainer.registerConfig(ModConfig.Type.COMMON, Config.SPEC);
     }
 
@@ -100,10 +123,11 @@ public class ExampleMod {
         // LOGGER.info("HELLO FROM COMMON SETUP");
 
         // if (Config.LOG_DIRT_BLOCK.getAsBoolean()) {
-        //     LOGGER.info("DIRT BLOCK >> {}", BuiltInRegistries.BLOCK.getKey(Blocks.DIRT));
+        // LOGGER.info("DIRT BLOCK >> {}", BuiltInRegistries.BLOCK.getKey(Blocks.DIRT));
         // }
 
-        // LOGGER.info("{}{}", Config.MAGIC_NUMBER_INTRODUCTION.get(), Config.MAGIC_NUMBER.getAsInt());
+        // LOGGER.info("{}{}", Config.MAGIC_NUMBER_INTRODUCTION.get(),
+        // Config.MAGIC_NUMBER.getAsInt());
 
         // Config.ITEM_STRINGS.get().forEach((item) -> LOGGER.info("ITEM >> {}", item));
     }
@@ -111,7 +135,7 @@ public class ExampleMod {
     // Add the example block item to the building blocks tab
     private void addCreative(BuildCreativeModeTabContentsEvent event) {
         // if (event.getTabKey() == CreativeModeTabs.BUILDING_BLOCKS) {
-        //     event.accept(EXAMPLE_BLOCK_ITEM);
+        // event.accept(EXAMPLE_BLOCK_ITEM);
         // }
 
         if (event.getTabKey() == CreativeModeTabs.INGREDIENTS) {
@@ -121,6 +145,18 @@ public class ExampleMod {
 
         if (event.getTabKey() == CreativeModeTabs.BUILDING_BLOCKS) {
             event.accept(ModBlocks.AZURITE_BLOCK);
+        }
+
+        if (event.getTabKey() == CreativeModeTabs.TOOLS_AND_UTILITIES) {
+            event.accept(ModItems.AZURITE_PICKAXE);
+            event.accept(ModItems.AZURITE_AXE);
+            event.accept(ModItems.AZURITE_SHOVEL);
+            event.accept(ModItems.AZURITE_HOE);
+        }
+
+        if (event.getTabKey() == CreativeModeTabs.COMBAT) {
+            event.accept(ModItems.AZURITE_SWORD);
+            event.accept(ModItems.AZURITE_SPEAR);
         }
     }