+++ src/main/java/com/example/examplemod/datagen/ModBlockTagsProvider.java
@@ -69,6 +69,14 @@ public class ModBlockTagsProvider extends BlockTagsProvider {
 
         tag(BlockTags.TRAPDOORS)
                 .add(ModBlocks.AZURITE_TRAPDOOR.getKey());
+
+        tag(ModTags.Blocks.NEEDS_AZURITE_TOOL)
+                .add(ModBlocks.MAGIC_BLOCK.getKey())
+                .addTag(BlockTags.NEEDS_IRON_TOOL);
+
+        tag(ModTags.Blocks.INCORRECT_FOR_AZURITE_TOOL)
+                .addTag(BlockTags.INCORRECT_FOR_IRON_TOOL)
+                .remove(ModTags.Blocks.NEEDS_AZURITE_TOOL);
     }
 
 }