+++ src/main/java/com/example/examplemod/item/ModToolTier.java
@@ -1,0 +1,10 @@
+package com.example.examplemod.item;
+
+import com.example.examplemod.tags.ModTags;
+
+import net.minecraft.world.item.ToolMaterial;
+
+public class ModToolTier {
+    public static final ToolMaterial AZURITE = new ToolMaterial(ModTags.Blocks.INCORRECT_FOR_AZURITE_TOOL, 1200,
+            3f, 3f, 22, ModTags.Items.AZURITE_REPAIRABLE);
+}